+++ src/Areas/Extension/Options/AzqrOptions.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+using AzureMcp.Options;
+
+namespace AzureMcp.Areas.Extension.Options;
+
+public class AzqrOptions : SubscriptionOptions;