+++ src/Areas/Extension/Options/ExtensionOptionDefinitions.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using AzureMcp.Models.Option;
+
 namespace AzureMcp.Areas.Extension.Options;
 
 public static class ExtensionOptionDefinitions
@@ -67,6 +69,17 @@ public static class ExtensionOptionDefinitions
                 Always run this command with learn=true and empty command on first run.
                 """
         )
+        {
+            IsRequired = false
+        };
+    }
+
+    public static class Azqr
+    {
+        public static readonly Option<string> OptionalResourceGroup = new(
+             $"--{OptionDefinitions.Common.ResourceGroupName}",
+             "The name of the Azure resource group. This is a logical container for Azure resources."
+         )
         {
             IsRequired = false
         };