+++ src/Commands/JsonSourceGenerationContext.cs
@@ -3,6 +3,7 @@
 
 using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
+using AzureMcp.Areas.Extension.Commands;
 using AzureMcp.Areas.Group.Commands;
 using AzureMcp.Commands;
 
@@ -14,6 +15,7 @@ namespace AzureMcp;
 [JsonSerializable(typeof(List<string>))]
 [JsonSerializable(typeof(List<JsonNode>))]
 [JsonSerializable(typeof(AzureCredentials))]
+[JsonSerializable(typeof(AzqrReportResult))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
 internal partial class JsonSourceGenerationContext : JsonSerializerContext
 {